Tweak FXAA + add light documentation for it
Sharpness is 8 by default which is fine, if not a tad too blurry for most games, but Minecraft is basically alias haven so I've decreased it to 6. 
edgeThreshold is 0.125 by default which again is fine for most games but in this case it's too much blurring for Minecraft as Minecraft has plainly visible textures with large surface areas that get blurred to oblivion at certain angles. In other words this higher value stops things from being aggressively blurred when they don't need to be. Also default value in sildur's vibrant shaders. 
edgeThresholdmin stops blurring in darker areas as aliasing is more difficult to see there; but that doesn't mean it isn't still visible. Performance is barely affected by this and gives a small boost in quality during night and in darker areas. 
subpixelremoval at lower or higher values increases aliasing and generates more noise, for some reason.
grain apparently makes FXAA ignore film grain but afaik this shader doesn't add any but no need to remove it in case you want to add that later.